개인과제 2차 작성_v1.3(App 중복 강제 형변환 수정)

+++ src/App.java
@@ -104,24 +104,24 @@ public class App {
 
                 switch (oper) {
                     case "+":
-                        result = cal.addOperator((int) firstNumber,secondNumber);
+                        result = cal.addOperator(firstNumber,secondNumber);
                         System.out.println(firstNumber + " + " + secondNumber + " = " + result);
                         break;
                     case "-":
-                        result = cal.subtractOperator((int) firstNumber,secondNumber);
+                        result = cal.subtractOperator(firstNumber,secondNumber);
                         System.out.println(firstNumber + " - " + secondNumber + " = " + result);
                         break;
                     case "*":
-                        result = cal.multiplyOperator((int) firstNumber,secondNumber);
+                        result = cal.multiplyOperator(firstNumber,secondNumber);
                         System.out.println(firstNumber + " * " + secondNumber + " = " + result);
                         break;
 
                     case "/":
-                        result = cal.divideOperator((int) firstNumber,secondNumber);
+                        result = cal.divideOperator(firstNumber,secondNumber);
                         System.out.println(firstNumber + " / " + secondNumber + " = " + result);
                         break;
                     case "%":
-                        result = cal.remainderOperator((int) firstNumber,secondNumber);
+                        result = cal.remainderOperator(firstNumber,secondNumber);
                         System.out.println(firstNumber + " % " + secondNumber + " = " + result);
                         break;
                     default:
